Script is now more stable + pause / resume functionality

+++ AdWatcher.py
@@ -16,6 +16,8 @@ getAdArea = True
 
 x = int(pyautogui.size()[0]/2 - 250)
 path = os.path.dirname(os.path.abspath(__file__))+"/".replace("\\", "/")
+blockScrolling = False
+numberOfClickedAds = 0
 
 if win32gui.FindWindow(None, "BlueStacks") != 0:  # if BlueStacks is running
     win = win32gui.FindWindow(None, "BlueStacks")
@@ -36,12 +38,20 @@ else:
 while 1:
     if win32gui.FindWindow(None, "BlueStacks") != 0:
         print("Starting...")
-        print("Note: Hold F5 to exit")
+        print("Hold F5 to exit\nHold F6 to pause\nPress F7 to resume")
         if keyboard.is_pressed("F5"):
             break
         while 1:
             if keyboard.is_pressed("F5"):  # Hold F5 to exit
                 break
+            if keyboard.is_pressed("F6"):  # Hold F6 to pause the script
+                print("Paused... \n  Press F7 to unpause")
+                while 1:
+                    if keyboard.is_pressed("F7"):
+                        print("Unpaused...")
+                        win32gui.SetForegroundWindow(win)
+                        win32gui.SetActiveWindow(win)
+                        break
 
             # The Area where a button will appear
             bluestacks_x = pyautogui.getWindowsWithTitle('BlueStacks')[
@@ -50,7 +60,7 @@ while 1:
             bluestacks_width = pyautogui.getWindowsWithTitle("BlueStacks")[
                 0].width - 375
             bluestacks_height = pyautogui.getWindowsWithTitle("BlueStacks")[
-                0].height - 685
+                0].height - 725
 
             # The Area of the entire BlueStacks window
             bluestacks_window_x = pyautogui.getWindowsWithTitle('BlueStacks')[
@@ -67,16 +77,18 @@ while 1:
 
             for i in cross:  # Go through all the images in the src/cross folder and compare them to the screen
                 cross_location = pyautogui.locateOnScreen(
-                    path+"src/cross/"+i, region=(bluestacks_x, bluestacks_y, bluestacks_width, bluestacks_height), confidence=0.75)
+                    path+"src/cross/"+i, region=(bluestacks_window_x, bluestacks_y, bluestacks_window_width, bluestacks_height), confidence=0.75)
                 if getAdArea:
                     if path+"ad_cap_area.png" in listdir(path):
                         os.remove(path+"ad_cap_area.png")
                     screenshot = pyautogui.screenshot(region=(
-                        bluestacks_x, bluestacks_y, bluestacks_width, bluestacks_height))
+                        bluestacks_x, bluestacks_y, bluestacks_window_width, bluestacks_height))
                     screenshot.save(path+"ad_cap_area.png")
                 if cross_location is not None:  # If the image is found on the screen then click the cross to end the Ad
+                    time.sleep(0.5)
                     pyautogui.click(cross_location)
                     print("Clicked cross")
+                    blockScrolling = False
                     if getDetectedCross:
                         if path+"cross_found.png" in listdir(path):
                             os.remove(path+"cross_found.png")
@@ -84,48 +96,52 @@ while 1:
                             cross_location[0], cross_location[1], cross_location[2], cross_location[3]))
                         screenshot.save(path +
                                         "cross_found.png")
-
                     time.sleep(1)
 
-            # Check if this was the last Ad
-            FinalAdCollected = pyautogui.locateOnScreen(path+"src/FinalAdCollected.png", confidence=0.9, region=(
-                bluestacks_window_x, bluestacks_window_y, bluestacks_window_width, bluestacks_window_height))
-            if path+"cap_area.png" in listdir(path):
-                os.remove(path+"cap_area.png")
-            screenshot = pyautogui.screenshot(region=(
-                bluestacks_window_x, bluestacks_window_y, bluestacks_window_width, bluestacks_window_height))
-            screenshot.save(path + "cap_area.png")
-            if FinalAdCollected is not None:  # If yes, then close the program
-                if getFinalAd:
-                    if path+"FinalAdCollected_found.png" in listdir(path):
-                        os.remove(path+"FinalAdCollected_found.png")
-                    screenshot = pyautogui.screenshot(region=(
-                        FinalAdCollected[0], FinalAdCollected[1], FinalAdCollected[2], FinalAdCollected[3]))
-                    screenshot.save(path + "FinalAdCollected_found"+".png")
-                print("Finished collecting ads")
-                time.sleep(3)
-                exit()
+            if numberOfClickedAds < 8:
 
-            # Check if there is another Ad to collect
-            NextAd = pyautogui.locateOnScreen(path+"src/WatchAd.png", confidence=0.5, region=(
-                bluestacks_window_x, bluestacks_window_y, bluestacks_window_width, bluestacks_window_height))
-            if NextAd is not None:
-                time.sleep(4)
-                pyautogui.click(NextAd)
-                print("Clicked Next Ad")
-                time.sleep(32)
-                if getNextAd:
-                    if path+"NextAd_found.png" in listdir(path):
-                        os.remove(path+"NextAd_found.png")
-                    screenshot = pyautogui.screenshot(region=(
-                        NextAd[0], NextAd[1], NextAd[2], NextAd[3]))
-                    screenshot.save(path + "NextAd_found"+".png")
-            else:  # If there is no next Ad then scroll down the screen and check again
-                pyautogui.moveTo(bluestacks_window_x + bluestacks_window_width/2,
-                                 bluestacks_window_y + bluestacks_window_height/2)
-                pyautogui.scroll(-100)
-                print("Could not find Next Ad, scrolling down")
-            time.sleep(2)
+                # Check if this was the last Ad
+                FinalAdCollected = pyautogui.locateOnScreen(path+"src/FinalAdCollected.png", confidence=0.975, region=(
+                    bluestacks_window_x, bluestacks_window_y, bluestacks_window_width, bluestacks_window_height))
+                if path+"cap_area.png" in listdir(path):
+                    os.remove(path+"cap_area.png")
+                screenshot = pyautogui.screenshot(region=(
+                    bluestacks_window_x, bluestacks_window_y, bluestacks_window_width, bluestacks_window_height))
+                screenshot.save(path + "cap_area.png")
+                if FinalAdCollected is not None:  # If yes, then close the program
+                    if getFinalAd:
+                        if path+"FinalAdCollected_found.png" in listdir(path):
+                            os.remove(path+"FinalAdCollected_found.png")
+                        screenshot = pyautogui.screenshot(region=(
+                            FinalAdCollected[0], FinalAdCollected[1], FinalAdCollected[2], FinalAdCollected[3]))
+                        screenshot.save(path + "FinalAdCollected_found"+".png")
+                    print("Finished collecting ads...\nExiting program now...")
+                    time.sleep(3)
+                    exit()
+
+                # Check if there is another Ad to collect
+                NextAd = pyautogui.locateOnScreen(path+"src/WatchAd.png", confidence=0.5, region=(
+                    bluestacks_window_x, bluestacks_window_y, bluestacks_window_width, bluestacks_window_height))
+                if NextAd is not None:
+                    time.sleep(1)
+                    pyautogui.click(NextAd)
+                    blockScrolling = True
+                    numberOfClickedAds = numberOfClickedAds + 1
+                    print("Clicked Next Ad:"+str(numberOfClickedAds)+"/8")
+                    time.sleep(35.5)
+                    if getNextAd:
+                        if path+"NextAd_found.png" in listdir(path):
+                            os.remove(path+"NextAd_found.png")
+                        screenshot = pyautogui.screenshot(region=(
+                            NextAd[0], NextAd[1], NextAd[2], NextAd[3]))
+                        screenshot.save(path + "NextAd_found"+".png")
+                # If there is no next Ad then scroll down the screen and check again
+                elif NextAd is None and blockScrolling == False:
+                    pyautogui.moveTo(bluestacks_window_x + bluestacks_window_width/2,
+                                     bluestacks_window_y + bluestacks_window_height/2)
+                    pyautogui.scroll(-100)
+                    print("Could not find Next Ad, scrolling down")
+                time.sleep(2)
     else:
         # If BlueStacks is not found exit the program
         if win32gui.FindWindow(None, "BlueStacks") == None:
